Update: updated all 'get user' methods to include the Place model. so, on log in, a user is returned with their places array included
