+++ back-end/server/api/users.js
@@ -19,7 +19,10 @@ router.get('/:id', async (req, res, next) => {
 router.put('/:id', async (req, res, next) => {
   try {
     const {imageURL, bio, email, name} = req.body;
-    const oldUser = await User.findByPk(req.params.id);
+    const oldUser = await User.findOne({
+      where: { id: req.params.id },
+      include: [{ model: Place, through: Snapshot }]
+    })
     const updatedUser = await oldUser.update({
       imageURL,
       bio,
